@@ -1,34 +1,34 @@
 {
-	"name": "n-anime-app",
-	"version": "0.1.0",
-	"private": true,
-	"scripts": {
-		"dev": "next dev --turbopack",
-		"tscheck": "tsc --noEmit",
-		"build": "next build --turbopack",
-		"start": "next start",
-		"lint": "next lint",
-		"format:fix": "prettier --write --ignore-path .gitignore .",
-		"preview": "next build && next start"
-	},
-	"dependencies": {
-		"@algolia/autocomplete-core": "1.19.4",
-		"babel-plugin-react-compiler": "^19.1.0-rc.1",
-		"clsx": "2.1.1",
-		"eslint-config-next": "^15.5.3",
-		"next": "^15.5.3",
-		"react": "^19.1.1",
-		"react-dom": "^19.1.1",
-		"react-hotkeys-hook": "5.1.0",
-		"sharp": "0.34.2",
-		"uuid": "^13.0.0"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "^2.2.4",
-		"@locator/babel-jsx": "^0.4.4",
-		"@types/node": "24.5.2",
-		"@types/react": "19.1.13",
-		"@types/react-dom": "19.1.9",
-		"typescript": "^5.9.2"
-	}
+  "name": "n-anime-app",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "tscheck": "bunx tsgo",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
+    "format:fix": "prettier --write --ignore-path .gitignore .",
+    "preview": "next build && next start"
+  },
+  "dependencies": {
+    "@algolia/autocomplete-core": "1.19.4",
+    "babel-plugin-react-compiler": "^1.0.0",
+    "clsx": "2.1.1",
+    "eslint-config-next": "^15.5.6",
+    "next": "^16.0.1",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "react-hotkeys-hook": "5.2.1",
+    "sharp": "0.34.2",
+    "uuid": "^13.0.0"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "^2.2.6",
+    "@locator/babel-jsx": "^0.4.4",
+    "@types/node": "24.8.1",
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2",
+    "typescript": "^5.9.3"
+  }
 }
