@@ -1,25 +1,28 @@
 import blurImage from '@/public/lights-blur.webp'
 import type { Anime } from '@/types'
 import Image from 'next/image'
-import { unstable_ViewTransition as ViewTransition } from 'react'
+import { ViewTransition } from 'react'
 import styles from './Anime.module.css'
 
 const PORTRAIT_ASPECT_RATIO = 9 / 12
 
 export async function AnimeAside({ anime }: { anime: Anime }) {
+	const imageSrc = anime.images?.coverImage || blurImage
+	const isExternalUrl = typeof imageSrc === 'string'
+
 	return (
 		<aside className={styles.aside}>
 			<ViewTransition name={`anime-image-${anime.animeId}`}>
 				<Image
-					src={anime.images?.coverImage || blurImage}
+					src={imageSrc}
 					alt={anime.title}
 					width={300}
 					height={300 / PORTRAIT_ASPECT_RATIO}
 					className={styles.asideImg}
 					priority
 					loading='eager'
-					blurDataURL={blurImage.src}
-					placeholder='blur'
+					placeholder={isExternalUrl ? 'empty' : 'blur'}
+					blurDataURL={!isExternalUrl ? blurImage.blurDataURL : undefined}
 				/>
 			</ViewTransition>
 			{anime.status && (
