@@ -1,4 +1,4 @@
-import { unstable_ViewTransition as ViewTransition } from 'react'
+import { ViewTransition } from 'react'
 import { BadgeList } from '../BadgeList'
 import styles from './Anime.module.css'
 
